Divide by byte.MaxValue instead of int.MaxValue

+++ Client/Main.cs
@@ -1124,7 +1124,7 @@ namespace GTANetwork
                 {
                     var ourVeh = NetEntityHandler.CreateObject(new Model(pair.Value.ModelHash), pair.Value.Position.ToVector(),
                         pair.Value.Rotation.ToVector(), false, pair.Key); // TODO: Make dynamic props work
-                    ourVeh.Opacity = pair.Value.Alpha / int.MaxValue;
+                    ourVeh.Opacity = pair.Value.Alpha / byte.MaxValue;
                 }
 
             if (map.Vehicles != null)
@@ -1180,7 +1180,7 @@ namespace GTANetwork
                     else
                         ourVeh.IsInvincible = true;
 
-                    ourVeh.Opacity = (float)pair.Value.Alpha / int.MaxValue;
+                    ourVeh.Opacity = pair.Value.Alpha / byte.MaxValue;
                 }
 
             if (map.Blips != null)
@@ -1233,7 +1233,7 @@ namespace GTANetwork
                         {
                             Function.Call(Hash.SET_PED_COMPONENT_VARIATION, ourPed, i, pair.Value.Props[i], pair.Value.Textures[i], 2);
                         }
-                        ourPed.Opacity = pair.Value.Alpha / int.MaxValue;
+                        ourPed.Opacity = pair.Value.Alpha / byte.MaxValue;
 
                         var ourSyncPed = Opponents.FirstOrDefault(op => op.Value.Character?.Handle == ourPed.Handle);
                         if (ourSyncPed.Value != null)
@@ -3230,7 +3230,7 @@ namespace GTANetwork
                 if (player.IsInAnyVehicle(false))
                 {
                     var veh = player.CurrentVehicle;
-                    veh.Opacity = (float)50 / int.MaxValue;
+                    veh.Opacity = (byte)50 / byte.MaxValue;
 
                     _debugSyncPed.VehiclePosition = veh.Position;
                     _debugSyncPed.VehicleRotation = veh.Rotation;
